- Upgraded to SUMO 0.13.0 APIs -- "srand" option replaced with "seed" -- STC.queryBounds() now reads all lanes and calculates the common boundary box -- removed deprecated RoadmapQuery.queryBounds() -- The following classes and methods use now double instead of floats: 	STC.changeEdgeTravelTime() 	STC.getEdgeTravelTime() 	Polygon.<init> 	RoadmapPosition 	ReadVehicleVarQuery.getRoadmapPositionFromResponse() 	RetrieveEdgeStateQuery.getGlobalTravelTime() -- Vehicle.changeRoute() renamed to setEdgeTravelTime() (for ongoing implementation of setRoute()) -- Vehicle.changeRouteMany() renamed to setEdgeTravelTimeMany() -- ChangeEdgeStateQuery.changeGlobalTravelTime() and ChangeVehicleStateQuery.changeEdgeTravelTime() now use double values - Query.queryAndVerify() now throws TraCIException()s with the string description returned from SUMO - ResponseMessage.<init> correctly handles error responses - added shape reading test - removed deprecated getMaxVehicleCount() - test/sumo_maps/variable_speed_signs upgraded to new 0.13.0 format

+++ src/it/polito/appeal/traci/protocol/RoadmapPosition.java
@@ -21,10 +21,10 @@ package it.polito.appeal.traci.protocol;
 
 public class RoadmapPosition {
 	public final String edgeID;
-	public final float pos;
+	public final double pos;
 	public final int laneID;
 	
-	public RoadmapPosition(String edgeID, float pos, int laneID) {
+	public RoadmapPosition(String edgeID, double pos, int laneID) {
 		this.edgeID = edgeID;
 		this.pos = pos;
 		this.laneID = laneID;
